contributors/rodneyrehm/incoming/css3-transitions/properties-value-inherit-001.html

<!DOCTYPE html>
<html>
    <head>
        <meta charset="utf-8">
    	<meta http-equiv="content-type" content="text/html; charset=utf-8">
        <title>CSS Test: transitioning inherited property values</title>
        <meta name="assert" content="Test checks that inhertied property values that are transitioned on a parent element don't start a transition">
        <link rel="help" href="http://www.w3.org/TR/css3-transitions/#starting">
        <link rel="author" title="Rodney Rehm" href="http://rodneyrehm.de/en/">
        <meta name="flags" content="dom invalid">
        
        <script src="/resources/testharness.js" type="text/javascript"></script>
        <script src="/resources/testharnessreport.js" type="text/javascript"></script>
        
        <script src="./support/helper.js" type="text/javascript"></script>
        <script src="./support/runAsyncHarness.js" type="text/javascript"></script>
        <script src="./support/generalTest.js" type="text/javascript"></script>
        <script src="./support/properties.js" type="text/javascript"></script>
    </head>
    <body>
        <!-- required by testharnessreport.js -->
        <div id="log"></div>
        <!-- elements used for testing -->
        <div id="fixture">
            <div id="container">
                <div id="transition">Hello World</div>
            </div>
        </div>

        <script>
            // see README.md for an explanation of how this test suite works
            // http://www.w3.org/TR/css3-transitions/#starting
            // Implementations also must not start a transition when the computed value changes because 
            // it is inherited (directly or indirectly) from another element that is transitioning the same property.
            
            // this test takes its time, give it 5 minutes to run
            setup({timeout: 300000});
            // init dom
            domFixture('#fixture');

            var tests = getPropertyTests();
            // for testing, limit to a couple of iterations
            // tests = tests.slice(10, 30);
            
            // missing support for requestAnimationFrame forces us to
            // run animations at a slower pace, so setTimeout has a
            // chance of being invoked before TransitionEnd
            var duration = getRequestAnimationFrame() ? '0.1s' : '0.5s';

            runAsyncHarness({
                // array of test data
                tests: tests,
                // milliseconds to wait before calling teardown and ending test
                duration: parseFloat(duration) * 1000,
                // prepare individual test
                setup: generalTest.setup,
                // cleanup after individual test
                teardown: generalTest.teardown,
                // run actual test, assertions can be used here!
                test: function(data, options) {
                    // clone and overwrite initial styles to be
                    // applied to #transition
                    var inherited = extend({}, data.from);
                    inherited[data.property] = 'inherit';
                    
                    var styles = {
                        // as we're testing inheritance, #fixture is our new parent
                        '#fixture': data.parentStyle,
                        // all styles including transition apply to to #container so they
                        // can inherit down to #transition
                        '#container': data.from,
                        '#container.to': data.to,
                        '#container.transition': {transition: data.property + ' ' + duration + ' linear 0s'},
                        // #transition only inherits and listens for transition events
                        '#transition': inherited,
                        '#transition.to' : {},
                        '#transition.transition' : {transition: data.property + ' ' + duration + ' linear 0s'}
                    };
                    
                    // prepare styles
                    generalTest.setStyle(data, styles);
                    // make sure values differ, if they don't, the property could most likely not be parsed
                    assert_not_equals(data.transition.from, data.transition.to, "initial and target values may not match");
                    // make sure value was properly inherited
                    assert_equals(data.transition.from, data.container.from, "inherited value must match origin");
                    // collect computed values during transition
                    generalTest.startValueCollection(data);
                    // kick off the transition
                    generalTest.startTransition(data);
                    
                    // make sure we didn't get the target value immediately.
                    // If we did, there wouldn't be a transition!
                    var current = data.transition.computedStyle(data.property);
                    assert_not_equals(current, data.transition.to, "must not be target value after start");
                },
                // run tests immediately before teardown, assertions can be used here!
                concluding: function(data) {
                    // make sure the parent triggered an event for the tested property only
                    data.test.step(generalTest.assertExpectedEventsFunc(data, 'container', data.property + ":" + duration));
                    // make sure we got no events on transition
                    data.test.step(generalTest.assertExpectedEventsFunc(data, 'transition', ""));
                    // make sure the property's value were neither initial nor target while transitioning
                    data.test.step(generalTest.assertIntermediateValuesFunc(data, 'transition'));
                },
                
                // called once all tests are done
                done: generalTest.done
            });
        </script>
    </body>
</html>
